+++ src-core/src/hardware/beyond/commands.rs
@@ -1,53 +1,53 @@
-use log::error;
+use tokio::sync::Mutex;
 
-use super::{super::HIDAPI, BEYOND_PID, BIGSCREEN_VID};
+use super::BEYOND_CONNECTED;
+
+lazy_static! {
+    static ref FAN_FORCED: Mutex<bool> = Mutex::new(false);
+}
 
 #[tauri::command]
 #[oyasumivr_macros::command_profiling]
-pub async fn bigscreen_beyond_set_brightness(brightness: u16) -> Result<(), String> {
-    if brightness >= 0x0400 {
-        error!(
-            "[Core] Attempted to set brightness of Bigscreen Beyond to an out of bounds value: {}",
-            brightness
-        );
-        return Err("OUT_OF_BOUNDS".to_string());
+pub async fn bigscreen_beyond_is_connected() -> bool {
+    let connected_guard = BEYOND_CONNECTED.lock().await;
+    *connected_guard
+}
+
+#[tauri::command]
+#[oyasumivr_macros::command_profiling]
+pub async fn bigscreen_beyond_set_brightness(
+    brightness: u16,
+    fan_safety: bool,
+) -> Result<(), String> {
+    // Handle fan safety
+    if fan_safety {
+        let should_force = brightness > 0x010a; // >100%
+        if should_force {
+            // Force fan to 100% if brightness is over 100%
+            match super::set_fan_speed(100).await {
+                Ok(_) => {
+                    let mut forced_guard = FAN_FORCED.lock().await;
+                    *forced_guard = true;
+                }
+                Err(e) => return Err(e),
+            }
+        } else {
+            // Restore the fan to 50% if it was forced before
+            let mut forced_guard = FAN_FORCED.lock().await;
+            if *forced_guard {
+                match super::set_fan_speed(50).await {
+                    Ok(_) => {
+                        *forced_guard = false;
+                    }
+                    Err(e) => return Err(e),
+                }
+                *forced_guard = false;
+            }
+        }
     }
-    let api_guard = HIDAPI.lock().await;
-    let api = match &*api_guard {
-        Some(a) => a,
-        None => {
-            error!("[Core] Attempted to call bigscreen_beyond_set_brightness, but hidapi is not initialized");
-            return Err("HIDAPI_NOT_INITIALIZED".to_string());
-        }
-    };
-    let device = match api.open(BIGSCREEN_VID, BEYOND_PID) {
-        Ok(d) => d,
-        Err(e) => {
-            error!("[Core] Could not open device for Bigscreen Beyond: {}", e);
-            return Err("DEVICE_NOT_FOUND".to_string());
-        }
-    };
-    let mut fan_percentage: u8 = 50;
-    if brightness > 266 {
-        fan_percentage = 100;
-    }
-    match device.send_feature_report(&[0, 0x46, fan_percentage as u8]) {
-        Ok(_) => {}
-        Err(e) => {
-            error!("[Core] Could not send data to Bigscreen Beyond: {}", e);
-            return Err("DEVICE_WRITE_ERROR".to_string());
-        }
-    };
-    match device.send_feature_report(&[
-        0,
-        0x49,
-        ((brightness >> 8) & 0xff) as u8,
-        (brightness & 0xff) as u8,
-    ]) {
+    // Set brightness
+    match super::set_brightness(brightness).await {
         Ok(_) => Ok(()),
-        Err(e) => {
-            error!("[Core] Could not send data to Bigscreen Beyond: {}", e);
-            return Err("DEVICE_WRITE_ERROR".to_string());
-        }
+        Err(e) => Err(e),
     }
 }